Fixed: type casting for wait_reply_message_ttl. For Bug in Carbon::addSeconds.

+++ core/config/app.php
@@ -141,5 +141,5 @@ return [
     |
     */
 
-    'wait_reply_message_ttl' => env('WAIT_REPLY_MESSAGE_TTL', 10),
+    'wait_reply_message_ttl' => (int)env('WAIT_REPLY_MESSAGE_TTL', 10),
 ];
